% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf_plot}
\alias{acf_plot}
\title{ACF Plots}
\usage{
acf_plot(
  x,
  lag.max = NULL,
  type = c("correlation", "covariance", "partial"),
  plot = FALSE,
  na.action = na.fail,
  demean = TRUE,
  ...,
  color = "steelblue",
  fill = "black",
  alpha = 0.7,
  data = NULL
)
}
\arguments{
\item{x}{a univariate or multivariate (not \code{ccf}) numeric time
    series object or a numeric vector or matrix, or an \code{"acf"} object.}

\item{lag.max}{maximum lag at which to calculate the acf.
    Default is \eqn{10\log_{10}(N/m)}{10*log10(N/m)} where \eqn{N} is the
    number of observations and \eqn{m} the number of series.  Will
    be automatically limited to one less than the number of observations
    in the series.}

\item{type}{character string giving the type of acf to be computed.
    Allowed values are
    \code{"correlation"} (the default), \code{"covariance"} or
    \code{"partial"}.  Will be partially matched.}

\item{plot}{A logical indicating whether the original base plot graphic should be displayed.
Generally this should be left to its default value of \code{FALSE}.}

\item{na.action}{function to be called to handle missing
    values. \code{na.pass} can be used.}

\item{demean}{logical.  Should the covariances be about the sample
    means?}

\item{...}{further arguments to be passed to \code{plot.acf}.}

\item{color}{color of reference lines (use "transparent" to hide)}

\item{fill}{fill color for ACF plot}

\item{alpha}{opacity of fill for ACF plot}
}
\description{
A wrapper around \code{stats::acf()} to make easy ACF plots in the ggplot2 system.
}
\examples{

acf_plot(rnorm(100))
acf_plot(rnorm(100), lag.max = 100, fill = "red", alpha = 0.3, color = "forestgreen")

}
\seealso{
\code{stats::acf()}
}
