% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Witmer-datadoc.R
\docType{data}
\name{Ancy}
\alias{Ancy}
\title{Plant growth experiment}
\format{A data frame with 3 variables:
\itemize{
\item \code{height}: height in cm after 14 days
\item \code{group}: \code{"Ancy"} or \code{"Control"}
\item \code{group_ind}: 1 for the control group and 2 for the Ancy group.
}}
\usage{
Ancy
}
\description{
A scientist treated 7 plants of the species \emph{Brassica campestris} with the
substance Ancymidol (Ancy) and compared them to 8 plants that were not treated.
Height (in cm) after 14 days of growth was measured for each plant.
}
\examples{
df_stats(height ~ group, data = Ancy)
}
\keyword{datasets}
