% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanfit.R
\name{as.mcmc.list.stanfit}
\alias{as.mcmc.list.stanfit}
\title{Convert stanfit Object to mcmc list Object}
\usage{
\method{as.mcmc}{list.stanfit}(x, ...)
}
\arguments{
\item{x}{a stanfit object}

\item{...}{additional arguments (ignored)}
}
\description{
Convert a stanfit object into an mcmc list object for use with coda and bayesplot tools.
}
