% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.R
\name{as.array.rjags}
\alias{as.array.rjags}
\alias{as.matrix.rjags}
\title{Convert objects to another type of object}
\usage{
\method{as.array}{rjags}(x, ...)

\method{as.matrix}{rjags}(x, ...)
}
\arguments{
\item{x}{an object to convert}

\item{...}{additional arguments}
}
\description{
Convert objects to another type of object
}
