% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta-dist.R
\name{betaParams}
\alias{betaParams}
\title{Compute parameters of a beta distribution}
\usage{
betaParams(shape1 = NULL, shape2 = NULL, mean = NULL, sd = NULL,
  concentration = NULL, mode = NULL, a = shape1, b = shape2,
  k = concentration, plot = FALSE, ...)
}
\arguments{
\item{shape1, a}{first shape parameter}

\item{shape2, b}{second shape paramter}

\item{mean}{mean of beta distribution}

\item{sd}{standard deviation of beta distribution}

\item{concentration, k}{concentration of beta distribution}

\item{mode}{mode of beta distribution}

\item{plot}{logical indicating whether a plot of the distribution
should be displayed.}

\item{...}{additional arguments passed to \code{\link[ggformula:gf_dist]{ggformula::gf_dist()}} to
modify the plot produced if \code{plot == TRUE}.}
}
\description{
Given two parameters which characterize a beta distribution,
compute the remaining parameters in the following set
\code{shape1} (\code{a}), \code{shape2} (\code{2}), \code{mean}, \code{mode}, \code{sd}, and \code{concentration} (\code{k}).
}
\examples{
betaParams(10, 5)
betaParams(mean = 0.8, concentration = 10)
betaParams(mean = 0.8, k = 10)
betaParams(mode = 0.8, concentration = 10)
betaParams(mode = 0.8, k = 10)
betaParams(mode = 0.8, k = 10, plot = TRUE, color = "red")
}
