% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoinExample.R
\name{CoinExample}
\alias{CoinExample}
\title{Example: Bayesian inference of a proportion}
\usage{
CoinExample(data = "HHTT", p = seq(0, 1, length.out = resolution + 1),
  prior = 1, results = c("plot", "data"), steps = FALSE,
  geom = geom_line, alpha = 0.7, resolution = 100, filter = TRUE,
  ...)
}
\arguments{
\item{data}{a vector of 0's and 1's or a character string with two characters representing
success/failure or head/tails from the bernoulli experiment.}

\item{p}{a vector of probabilites to update.}

\item{prior}{a vector of prior values at each probability in \code{p} or a constant.}

\item{results}{either \code{"plot"} or \code{"data"} to indicate what information is returned.}

\item{steps}{a logical indicating wether the process should be shown step-by-step or
for all the data at once.}

\item{geom}{a geom used for plotting.  Good choices include
\code{\link[ggplot2:geom_line]{ggplot2::geom_line()}},
\code{\link[ggplot2:geom_col]{ggplot2::geom_col()}},
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_area]{ggplot2::geom_area()}}}

\item{alpha}{opacity used for the plot}

\item{resolution}{if \code{p} is not specified, \code{resolution} an be used to create an equally
spaced grid of \code{resolution + 1} values from 0 to 1.}

\item{...}{additional arguments passed to the geom.}
}
\description{
This intended as a black-box example with a small data set to illustrate
the results Bayesian updating.
}
\examples{
CoinExample("HHTTH", geom = geom_col, p = seq(0, 1, by = 0.25), results = "data")
CoinExample("HHTTH",
  p = seq(0, 1, by = 0.25), prior = c(.1, .2 , .4, .2, .1),
  geom = geom_col, alpha = 0.5)
CoinExample("HHTTH", geom = geom_col, p = seq(0, 1, by = 0.25), steps = TRUE, results = "data")
CoinExample("HHTTH", p = seq(0, 1, by = 0.25), alpha = 0.9, steps = TRUE)
CoinExample("HHTTH", p = seq(0, 1, by = 0.25), alpha = 0.5, steps = TRUE,
  geom = geom_col, filter = n \%in\% c(4,5))
CoinExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE)
CoinExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE)
CoinExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE, filter = n \%in\% c(1, 5, 10))
CoinExample("HHHHHHHHHHHH", alpha = 0.9)
CoinExample("HHHHHHHHHHHH", alpha = 0.9, steps = TRUE)
}
