% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDAplots.R
\name{diagMCMC}
\alias{diagMCMC}
\alias{diag_mcmc}
\title{Diagnostic Plots for MCMC samples}
\usage{
diagMCMC(object, parName = varnames(object)[1])

diag_mcmc(object, parName = varnames(object)[1])
}
\arguments{
\item{object}{an MCMC object as per the coda package}

\item{parName}{an alias for pars}

\item{pars}{the name of a parameter for which diagnostics are displayed}
}
\description{
This function mimics \code{diagMCMC()} from Kruschke's \emph{Doing Bayesian Data Analysis}.
The options to save the plot to a file have been removed since there are other
general purpose ways of doing this, and it isn't needed if you are working in
R Markdown, which is recommended for reporting.
}
