% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamma-dist.R
\name{gamma_params}
\alias{gamma_params}
\title{Compute parameters of a gamma distribution}
\usage{
gamma_params(shape = NULL, rate = NULL, scale = NULL, mean = NULL,
  sd = NULL, mode = NULL, plot = FALSE, ...)
}
\arguments{
\item{shape, }{shape parameter}

\item{rate, }{rateparamter}

\item{mean}{mean of gamma distribution}

\item{sd}{standard deviation of gamma distribution}

\item{mode}{mode of gamma distribution}

\item{plot}{logical indicating whether a plot of the distribution
should be displayed.}

\item{...}{additional arguments passed to \code{\link[ggformula:gf_dist]{ggformula::gf_dist()}} to
modify the plot produced if \code{plot == TRUE}.}

\item{concentration, k}{concentration of gamma distribution}
}
\description{
Given two parameters which characterize a gamma distribution,
compute the remaining parameters in the following set
\code{shape}, \code{rate}, \code{scale}, \code{mean}, \code{mode}, \code{sd}.
}
\examples{
gamma_params(10, 5)
gamma_params(10, scale = 5)
gamma_params(mean = 10, sd = 3)
gamma_params(mode = 10, sd = 3, plot = TRUE)
gamma_params(mean = 1, sd = 10, plot = TRUE)
}
