% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{gg_dag}
\alias{gg_dag}
\title{Plot DAGs simply}
\usage{
gg_dag(dag, size = 7, highlight = "U")
}
\arguments{
\item{dag}{a DAG}

\item{size}{size of the nodes}

\item{highlight}{a vector of names of nodes to display in a highlighted color}
}
\description{
Create a simple ggplot displaying a DAG created with \code{dagitt::dagitty}
or \code{ggdag::dagify()}.
}
\examples{
ggdag::dagify(A ~ B + C, B ~ C, D ~ C) \%>\%
 gg_dag(highlight = c("B", "C"), size = 9 )
}
