% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjags.R
\name{neff}
\alias{neff}
\title{Effective Sample Size for MCMC objects}
\usage{
neff(object, ...)
}
\arguments{
\item{object}{an object containing information about MCMC samples}

\item{...}{additional arguments passed to specific methods}

\item{regex_pars}{a vector of regular expressions matching the parameters for which
effective sample size will be returned.}
}
\description{
Extract effective sample size information from MCMC objects.
\code{neff()} is a generic function. See below for objects that
\code{neff()} knows how to handle.
}
