% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peek.R
\name{peek}
\alias{peek}
\title{Perform side effect}
\usage{
peek(object, fun = function(x, ...) x, ...)
}
\arguments{
\item{object}{An R object}

\item{fun}{a function that can be applied to \code{object}}

\item{...}{additional arguments to \code{fun}}
}
\description{
Perform side effect
}
\examples{

if (require(dplyr)) {
  iris \%>\%
  peek(slice_sample, n = 3) \%>\%
  group_by(Species) \%>\%
  summarise(n = n())
}

}
