% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDAplots.R
\name{plotPost}
\alias{plotPost}
\alias{plot_post}
\title{Plot posterior samples}
\usage{
plotPost(samples, center = c("mode", "median", "mean"),
  cenTend = center, comparison_value = NULL,
  compVal = comparison_value, ROPE = NULL, hdi_prob = 0.95,
  credMass = hdi_prob, hdi_text_place = 0.7,
  HDItextPlace = hdi_text_place, quietly = FALSE,
  xlab = "Param. Val.", xlim = NULL, yaxt = "n", ylab = "",
  main = "", cex = 1.4, cex.lab = 1.5, col = "skyblue",
  border = "white", showCurve = FALSE, breaks = NULL, ...)

plot_post(samples, center = c("mode", "median", "mean"),
  cenTend = center, comparison_value = NULL,
  compVal = comparison_value, ROPE = NULL, hdi_prob = 0.95,
  credMass = hdi_prob, hdi_text_place = 0.7,
  HDItextPlace = hdi_text_place, quietly = FALSE,
  xlab = "Param. Val.", xlim = NULL, yaxt = "n", ylab = "",
  main = "", cex = 1.4, cex.lab = 1.5, col = "skyblue",
  border = "white", showCurve = FALSE, breaks = NULL, ...)
}
\arguments{
\item{samples}{a vector of (posterior) sampled values}

\item{cenTend}{measure of central tendency to use.
One of \code{"mode"}, \code{"median"} or \code{"mean"}.}

\item{comparison_value, }{compVal number for "comparison value"}

\item{ROPE}{a vector of length 2 giving the ends of the region of practical equivalence}

\item{hdi_prob, credMass}{probability for HDI interval}

\item{hdi_text_place, HDItextPlace}{location of HDI text on plot.}

\item{quietly}{if TRUE, return summary invisibly.}

\item{xlab, xlim, yaxt, ylab, main, cex, cex.lab, col, border, showCurve, breaks}{arguments for
\code{\link[graphics:hist]{graphics::hist()}}}

\item{...}{additional arguments passed along to \code{\link[graphics:hist]{graphics::hist()}}}
}
\description{
This function mimics \code{plotPost()} from Kruschke's \emph{Doing Bayesian Data Analysis}
and provides a highly customized plot of a vector of posterior samples for
a parameter.
The options to save the plot to a file have been removed since there are other
general purpose ways of doing this, and it isn't needed if you are working in
R Markdown, which is recommended for reporting.
}
\examples{
samples <- rnorm(2000)    # fake a posterior sample
plot_post(samples)
plot_post(samples, hdi_prob = 0.90)
plot_post(samples, hdi_prob = 0.90, ROPE = c(-0.2, 0.2), comparison_value = 2)
plot_post(samples, hdi_prob = 0.90, ROPE = c(-0.2, 0.2), compVal = 2)

}
