% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{posterior}
\alias{posterior}
\title{Extract posterior samples as a tidy data frame}
\usage{
posterior(object, ...)
}
\arguments{
\item{object}{An object produced by a bayesian fitting function like
\code{\link[R2jags:jags]{R2jags::jags()}}.}
}
\description{
Extract posterior samples as a tidy data frame
}
