% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_calc.R
\name{posterior_calc}
\alias{posterior_calc}
\alias{posterior_calc.default}
\alias{posterior_calc.data.frame}
\title{Make posterior cacluations}
\usage{
posterior_calc(object, ...)

\method{posterior_calc}{default}(object, formula, data = NULL,
  draws = NULL, data.frame = FALSE, ...)

\method{posterior_calc}{data.frame}(object, formula, data = NULL,
  draws = NULL, data.frame = FALSE, ...)
}
\arguments{
\item{object}{An object from which posterior calculations are made.}

\item{...}{Additional arguments, currently ignored.}

\item{formula}{A formula describing the quantity to be calculated.  The rhs of the
formula is evaluated using one posterior sample of the parameters and \code{data}.
The lhs, if it exists, is used to name the resulting column when \code{data.frame = TRUE}.}

\item{data}{Additional data involved in the computation. This may be the original
data used to fit the model or counter-factual data.}

\item{draws}{The number of draws to make from the posterior distribution.
Sampling is with replacement if \code{draws} is larger than the number of
posterior samples in \code{object}.}

\item{data.frame}{A logical indicating whtehr the results should be returned
as a data frame (TRUE) or a matrix (FALSE).}
}
\value{
A matrix with \code{draws} rows or data frame with three columns.
}
\description{
A generic and some methods for creating posterior calculations.
The intended use is to prepare data for use with bayesplot ppc plots.
The default method works for any object for which \code{posterior()} can
create a data frame of posterior samples.
}
