% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjags.R
\name{print.summary.rjags}
\alias{print.summary.rjags}
\title{Print summary.rjags objects}
\usage{
\method{print}{summary.rjags}(x, digits = NULL, ...)
}
\arguments{
\item{x}{a summary.rjags objejct}
}
\description{
Print summary.rjags objects
}
