% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropExample.R
\name{PropExample}
\alias{PropExample}
\title{Example: Bayesian inference of a proportion}
\usage{
PropExample(data = "HHTT", p = seq(0, 1, length.out = resolution + 1),
  prior = 1, results = c("plot", "data"), steps = FALSE,
  geom = geom_line, alpha = 0.7, resolution = 100, filter = TRUE,
  ...)
}
\arguments{
\item{data}{a vector of 0's and 1's or a character string with two characters representing
success/failure or head/tails from the bernoulli experiment.}

\item{p}{a vector of probabilites to update.}

\item{prior}{a vector of prior values at each probability in \code{p} or a constant.}

\item{results}{either \code{"plot"} or \code{"data"} to indicate what information is returned.}

\item{steps}{a logical indicating wether the process should be shown step-by-step or
for all the data at once.}

\item{geom}{a geom used for plotting.  Good choices include
\code{\link[ggplot2:geom_line]{ggplot2::geom_line()}},
\code{\link[ggplot2:geom_col]{ggplot2::geom_col()}},
\code{\link[ggplot2:geom_point]{ggplot2::geom_point()}},
\code{\link[ggplot2:geom_area]{ggplot2::geom_area()}}}

\item{alpha}{opacity used for the plot}

\item{resolution}{if \code{p} is not specified, \code{resolution} an be used to create an equally
spaced grid of \code{resolution + 1} values from 0 to 1.}

\item{...}{additional arguments passed to the geom.}
}
\description{
This intended as a black-box example with a small data set to illustrate
the results of Bayesian updating.  The model is a Bernoulli model with
fixed proportion over all tosses of the coin.  Prior, data, and grid resolution
can be supplied by the user. See examples.
}
\examples{
PropExample("HHTTH", geom = geom_col, p = seq(0, 1, by = 0.25), results = "data")
PropExample("HHTTH",
  p = seq(0, 1, by = 0.25), prior = c(.1, .2 , .4, .2, .1),
  geom = geom_col, alpha = 0.5)
PropExample("SSFFS", geom = geom_col, p = seq(0, 1, by = 0.25), steps = TRUE, results = "data")
PropExample("SSFFS", p = seq(0, 1, by = 0.25), alpha = 0.9, steps = TRUE)
PropExample("SSFFS", p = seq(0, 1, by = 0.25), alpha = 0.5, steps = TRUE,
  geom = geom_col, filter = n \%in\% c(4,5))
PropExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE)
PropExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE)
PropExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE, filter = n \%in\% c(1, 5, 10))
PropExample("HHTTHTHHTHHH", alpha = 0.9, steps = TRUE, filter = n == 10)
# same posterior as in the previous example -- order doesn't matter
PropExample("TTTTHHHHHHHH", alpha = 0.9, steps = TRUE, filter = n == 10)
PropExample("HHHHHHHHHHHH", alpha = 0.9)
PropExample("HHHHHHHHHHHH", alpha = 0.9, steps = TRUE)
PropExample(rbinom(12, 1, 0.25), alpha = 0.9, steps = TRUE)
}
