% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{stanfit}
\alias{stanfit}
\title{Extract Stan fit from wrapper object}
\usage{
stanfit(object, ...)
}
\arguments{
\item{object}{An object from which to extract a stanfit object.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
The default method with search a list and return an element that inherits from class "stanfit",
if one exists.
}
