% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rjags.R
\name{summary.rjags}
\alias{summary.rjags}
\title{Summarize an rjags object}
\usage{
\method{summary}{rjags}(object, digits = 3, intervals = c(0.025, 0.25,
  0.5, 0.75, 0.975), ...)
}
\arguments{
\item{object}{and rjags object}

\item{digits}{preferred number of digits to display}

\item{intervals}{quantiles of posterior to diplay}

\item{...}{additional arguments}
}
\description{
Summarize an rjags object
}
