% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_one.R
\name{zero_one}
\alias{zero_one}
\title{Convert vector to 0's and 1's}
\usage{
zero_one(x, one = NULL)
}
\arguments{
\item{x}{a vector}

\item{one}{A vector of values to be coded as 1.  All other values will be coded as 0.
If \code{NULL}, a defult choice is made.}
}
\description{
JAGS and Stan generally require the dichotymous data be coded as 0's and 1's. This
function converts other data fromats to 0's and 1's.
}
\examples{
zero_one(iris$Species)
}
