% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick-bern-beta.R
\name{quick_bern_beta}
\alias{quick_bern_beta}
\title{Quick determination of posterior from a beta prior and Bernoulli likelihood}
\usage{
quick_bern_beta(x, n, ...)
}
\arguments{
\item{x}{number of successes in data}

\item{n}{number of trials in data}

\item{...}{prameters describing the beta prior.  Uses \code{\link[=beta_params]{beta_params()}}.}
}
\description{
Quick determination of posterior from a beta prior and Bernoulli likelihood
}
\examples{
quick_bern_beta(x = 4, n = 5, a = 2, b = 4)
quick_bern_beta(x = 4, n = 5, mode = 0.3, concentration = 10)
}
