% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student_t.R
\name{rstudent_t}
\alias{rstudent_t}
\title{Shifted and scaled t distributions}
\usage{
rstudent_t(n, df, mean = 0, sd = 1)
}
\arguments{
\item{n}{Number of random draws to perform.}

\item{df, mean, sd}{Degrees of freedom, mean, and standard deviation of distribution}
}
\value{
A vector of random draws from the appropriate t distribution(s).
}
\description{
The functions are designed for compatibility with Stan's student t distribution
which has three parameters: degrees of freedom, mean, and standard deviation.
}
